Move _publish_gripping_point_debug_data to gripping_points_tools.py
